requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
pyserial>=3.5
dearpygui>=1.6.2
joblib>=1.1.0
pytest>=6.2.5
biosppy>=0.7.0
pyopengl>=3.1.5
tqdm>=4.62.0
pillow>=8.3.0
pyqt5>=5.15.0
pygame>=2.0.0
